# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: sg-gateway API
  version: 0.1.0
  description: |
    DCDeepTech Singapore control-plane gateway.

    All requests require a Bearer API key in the `Authorization` header.
    Keys are issued via the `/admin/api-keys` endpoints and have the format `sk-dcdt-*`.

    **Cross-border policy** is enforced server-side on every inference request:
    - `PUBLIC` / `LOW_RISK` → routed to Chongqing (CQ) vLLM
    - `PERSONAL` → redacted, routed to Singapore LiteLLM
    - `HIGH_RISK` → rejected with HTTP 451

servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.dcdeeptech.sg
    description: Production (Singapore)

tags:
  - name: inference
    description: OpenAI-compatible inference endpoint
  - name: admin
    description: API key management (requires admin:keys scope)
  - name: compliance
    description: PDPA compliance (proxied from sg-compliance service)

paths:
  /health:
    get:
      summary: Health check
      operationId: health
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /v1/chat/completions:
    post:
      summary: Chat completions (OpenAI-compatible)
      operationId: chatCompletions
      tags: [inference]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Completion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "451":
          description: Unavailable for legal reasons — HIGH_RISK data detected, transfer blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /admin/api-keys:
    get:
      summary: List API keys
      operationId: listApiKeys
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of API keys (prefix only — full key never returned after creation)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      summary: Create API key
      operationId: createApiKey
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKeyRequest"
      responses:
        "201":
          description: Key created — full key returned ONCE only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateKeyResponse"

  /admin/api-keys/{key_id}/disable:
    post:
      summary: Disable an API key (reversible)
      operationId: disableApiKey
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key disabled

  /admin/api-keys/{key_id}/revoke:
    post:
      summary: Revoke an API key (irreversible)
      operationId: revokeApiKey
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked permanently

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key in format `sk-dcdt-*`

  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: API key does not have the required scope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  schemas:
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          example: deepseek-r1-8b
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1.0
        max_tokens:
          type: integer

    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          example: chat.completion
        created:
          type: integer
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        x_dcdt_policy:
          type: object
          description: DCDeepTech routing metadata
          properties:
            data_class:
              type: string
              enum: [PUBLIC, LOW_RISK, PERSONAL, HIGH_RISK]
            destination:
              type: string
              enum: [sg, cq]
            redacted:
              type: boolean

    ApiKey:
      type: object
      properties:
        key_id:
          type: string
          example: kdcdt_a1b2c3d4
        key_prefix:
          type: string
          example: "sk-dcdt-a1b2c3d4e5f6g7h8..."
        tenant_id:
          type: string
        scopes:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, disabled, revoked]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          nullable: true

    CreateKeyRequest:
      type: object
      required: [tenant_id, scopes]
      properties:
        tenant_id:
          type: string
        scopes:
          type: array
          items:
            type: string
            enum: [chat:completions, admin:keys, admin:tenants, internal:routing]
        expires_in_days:
          type: integer
          nullable: true
        description:
          type: string

    CreateKeyResponse:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          properties:
            key:
              type: string
              description: Full API key — returned ONCE only. Store immediately.
              example: sk-dcdt-a1b2c3d4e5f6...

    Error:
      type: object
      properties:
        detail:
          type: string